speed of map and float coordinate
Add change speed of map when character move too fast. Change type of coordinate of Lane and Enemy from int to float

+++ src/Class/Core.h
@@ -7,7 +7,7 @@
 class Core {
 public:
     Core();
-    double getSpeedMultiplier();
+    float getSpeedMultiplier();
     bool detectCollision();
     bool detectBlockMovement(int direction);
     void update(float dt);
